requirements.txt
fschat==0.2.23
pandas 
seaborn 
matplotlib 
numpy 
datasets 
evaluate
anthropic 
google-generativeai 
openai
SentencePiece
accelerate
google-cloud-aiplatform
langchain-community
faiss-cpu
langchain-huggingface
sentence-transformers
transformers